requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
cron-descriptor==1.4.3
Django==5.0.7
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.7
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
rest-framework-simplejwt==0.0.2
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
pytest
